adds hash routing for future improvements
