+++ package.json
@@ -28,6 +28,7 @@
     "jest": "^24.8.0",
     "jest-cli": "^24.8.0",
     "npm-run-all": "^4.1.5",
+    "page": "^1.11.6",
     "prettier": "2.0.5",
     "prettier-plugin-svelte": "1.1.0",
     "rollup": "2.16.1",
@@ -72,9 +73,6 @@
     "d3-shape": "^1.3.5",
     "d3-time": "^1.0.11",
     "d3-time-format": "^2.1.3",
-    "express": "^4.17.1",
-    "popper.js": "^1.15.0",
-    "sqlite3": "^4.0.9",
-    "tooltip.js": "^1.3.2"
+    "express": "^4.17.1"
   }
 }
