+++ src/App.svelte
@@ -1,25 +1,18 @@
 <script>
-  import { onMount } from 'svelte';
+  import page from 'page';
   import Sidebar from "./components/Sidebar.svelte";
-  import Explore from "./components/Explore.svelte";
-  import LoadingSpinner from './components/LoadingSpinner.svelte';
-  import ErrorMessage from './components/ErrorMessage.svelte';
-  import { createRequestCache, storeToQuery } from './stores/cache';
+  import routes from './routes/routes';
   import { store } from './stores/store';
+  import { storeToQuery } from './stores/cache';
 
-  const cache = createRequestCache();
-
-  function updateQueryString(value) {
-    if (history.pushState) {
-      const newurl =
-        window.location.protocol +
-        "//" +
-        window.location.host +
-        window.location.pathname +
-        `?${value}`;
-      window.history.pushState({ path: newurl }, "", newurl);
+  function updateQueryString(query) {
+    if (window.history.replaceState) {
+      const newURL = `${window.location.origin}${window.location.pathname}?${query}`;
+      window.history.replaceState(null, null, newURL);
     }
   }
+
+
 
   let mounted = false;
   onMount(() => { mounted = true });
@@ -27,6 +20,21 @@
   $: if (mounted) {
     updateQueryString(storeToQuery($store));
   }
+
+  let bodyComponent;
+  let sidebarComponent;
+
+  function useComponent(thisView = undefined) {
+    return ({params}) => {
+      const currentView = params.view;
+      let view = thisView || currentView;
+      bodyComponent = routes[view].body;
+      sidebarComponent = routes[view].sidebar;
+    }
+  }
+  page('/', useComponent('explore'));
+  page('/:view', useComponent());
+  page({ hashbang: true });
 
 </script>
 
@@ -39,12 +47,8 @@
 </style>
 
 <div>
-  <Sidebar />
-  {#await $cache}
-    <LoadingSpinner />
-  {:then value}
-    <Explore data={value} />
-  {:catch err}
-    <ErrorMessage message={err.message} />
-  {/await}
+  <Sidebar>
+    <svelte:component this={sidebarComponent} />
+  </Sidebar>
+  <svelte:component this={bodyComponent} />
 </div>
