+++ src/routes/ytd/Sidebar.svelte
@@ -1,0 +1,3 @@
+<div>
+ytd.
+</div>
